+++ semester2/hw02/hw02_task01/List/List.h
@@ -6,12 +6,12 @@ class List
 public:
     virtual ~List() {}
 
-    virtual void insert(int position, T item) = 0;
-    virtual void remove(int position) = 0;
-    virtual T & at(int position) = 0;
+    virtual void insert(size_t position, T item) = 0;
+    virtual void remove(size_t position) = 0;
+    virtual T & at(size_t position) = 0;
 
     virtual int find(T item) const = 0;
-    virtual int length() const = 0;
+    virtual size_t length() const = 0;
 
     static const int itemNotFound = -1;
 };
